backend/requirements.txt
fastapi
uvicorn
gunicorn
pymongo
dnspython
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
certifi
python-multipart
pydantic-settings
